279: Move fallbacks
