+++ frontend/src/routes/page.tsx
@@ -1,7 +1,7 @@
 import Demo from '../components/Demo';
 import { Center, Container, Title } from '@mantine/core';
 import { ErrorBoundary } from 'react-error-boundary';
-import { PageFallback } from './fallbacks/PageFallback.tsx';
+import { PageFallback } from './PageFallback.tsx';
 
 const Root = () => {
   return (
