+++ frontend/src/routes/settings/page.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@mantine/core';
 import { ErrorBoundary } from 'react-error-boundary';
-import { PageFallback } from '../fallbacks/PageFallback.tsx';
+import { PageFallback } from '../PageFallback.tsx';
 
 const Settings = () => {
   return <Box>Work in progress</Box>;
